+++ tests/igs/test_extract.py
@@ -5,15 +5,20 @@ import pytest
 from mex.common.ldap.models import LDAPActor
 from mex.common.models import AccessPlatformMapping, ResourceMapping
 from mex.extractors.igs.extract import (
+    extract_endpoint_counts,
     extract_igs_schemas,
     extract_ldap_actors_by_mail,
 )
+from mex.extractors.igs.model import IGSSchema
 
 
 @pytest.mark.usefixtures("mocked_igs")
 def test_extract_igs_schemas() -> None:
     schemas = extract_igs_schemas()
-    assert schemas["Pathogen"].model_dump() == {"enum": ["PATHOGEN"], "type": "string"}
+    assert schemas["igsmodels__enums__Pathogen"].model_dump() == {
+        "enum": ["PATHOGEN"],
+        "type": "string",
+    }
     assert schemas["schemaCreation"].model_dump() == {
         "properties": {
             "schemas": {
@@ -22,6 +27,18 @@ def test_extract_igs_schemas() -> None:
                 "title": "test_title",
             },
         }
+    }
+
+
+@pytest.mark.usefixtures("mocked_igs")
+def test_extract_endpoint_counts(
+    igs_resource_mapping: ResourceMapping, igs_schemas: dict[str, IGSSchema]
+) -> None:
+    endpoint_counts = extract_endpoint_counts(igs_resource_mapping, igs_schemas)
+    assert endpoint_counts == {
+        "/genomes/count": "7",
+        "pathogen_PATHOGEN": "4",
+        "/uploads/count": "7",
     }