+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - prepare testing against a mocked ldap container
 - primary source helper function: add caching
 - implemented asset checks for the `x_items_less_than` rule
+- extract counts from IGS API
 
 ### Changes